requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
pydantic-settings
passlib
bcrypt
python-jose
python-dotenv
email-validator
python-multipart